added contextual link to US laws exclusion
As a public policy that is not explicitly stated in the Copyright Act: http://en.wikipedia.org/wiki/Edict_of_government. I thought it was good idea to link to the Copyright Office's Guidelines stating the Edicts of Government exclusion. This is not the official gov link though, which is not promising: http://www.copyright.gov/compendium/